Add check for duplicate arcs in `generate` method
Now `generate` will check that a newly generated arc/column is unique in the matrix, preventing multiple repeat arcs.

If it is a duplicate, `generate` will attempt to generate another new unique arc/column, and will continue to do so up to MAX_RETRIES times. If it cannot find a new unique one, the column will simply consist of all zeros.

In the future, we might just remove these columns of zeros which would result in a smaller matrix than expected.